refactor: move professional naming implementations behind canonical modules
