+++ src/sdetkit/platform_readiness_kickoff.py
@@ -1,17 +1,9 @@
-"""Compatibility alias for the professional naming migration."""
+"""Compatibility wrapper for historical `sdetkit.platform_readiness_kickoff` imports."""
 
 from __future__ import annotations
 
-from ._compat_alias import alias_dir as _alias_dir
-from ._compat_alias import alias_getattr as _alias_getattr
-from ._compat_alias import export_module as _export_module
+from importlib import import_module as _import_module
 
-_TARGET = _export_module("sdetkit.phase3_kickoff", globals())
-
-
-def __getattr__(name: str) -> object:
-    return _alias_getattr(_TARGET, name)
-
-
-def __dir__() -> list[str]:
-    return _alias_dir(globals(), _TARGET)
+_IMPL = _import_module("sdetkit.phases.platform_readiness_kickoff")
+__all__ = getattr(_IMPL, "__all__", [name for name in dir(_IMPL) if not name.startswith("__")])
+globals().update({name: getattr(_IMPL, name) for name in __all__})
