+++ src/sdetkit/release_readiness_utilities.py
@@ -1,17 +1,21 @@
-"""Compatibility alias for the professional naming migration."""
-
 from __future__ import annotations
 
-from ._compat_alias import alias_dir as _alias_dir
-from ._compat_alias import alias_getattr as _alias_getattr
-from ._compat_alias import export_module as _export_module
-
-_TARGET = _export_module("sdetkit.phase2_utilities", globals())
+import datetime as dt
 
 
-def __getattr__(name: str) -> object:
-    return _alias_getattr(_TARGET, name)
+def parse_check_csv(value: str | None) -> list[str]:
+    if value is None:
+        return []
+    out: list[str] = []
+    for part in value.split(","):
+        item = part.strip()
+        if item:
+            out.append(item)
+    return out
 
 
-def __dir__() -> list[str]:
-    return _alias_dir(globals(), _TARGET)
+def parse_iso_date(raw: str, *, field: str) -> dt.date:
+    try:
+        return dt.date.fromisoformat(raw)
+    except ValueError as exc:
+        raise ValueError(f"{field} must be ISO date YYYY-MM-DD") from exc
